+++ Sources/SwiftTUIViews/Controls/ProgressViewStyles.swift
@@ -110,7 +110,7 @@ public struct AnyProgressViewStyle: Sendable, CustomStringConvertible, CustomDeb
 
   public init<S: ProgressViewStyle>(_ style: S) {
     snapshotLabel = style.snapshotLabel
-    box = ConcreteAnyProgressViewStyleBox(style: style)
+    box = ConcreteStyleBox(style: style)
   }
 
   public var description: String { snapshotLabel }
@@ -192,29 +192,22 @@ extension ProgressViewStyle where Self == CircularProgressViewStyle {
 
 extension CircularProgressViewStyle: ReuseTransparentStyle {}
 
-private protocol AnyProgressViewStyleBox: Sendable {
-  func isEqualForReuse(to other: any AnyProgressViewStyleBox) -> Bool
+private protocol AnyProgressViewStyleBox: AnyStyleBox {
 
   @MainActor
   func resolveBody(configuration: ProgressViewStyleConfiguration, in context: ResolveContext)
     -> ResolvedNode
 }
 
-private struct ConcreteAnyProgressViewStyleBox<S: ProgressViewStyle>: AnyProgressViewStyleBox {
-  let style: S
-
-  func isEqualForReuse(to other: any AnyProgressViewStyleBox) -> Bool {
-    guard let other = other as? Self else { return false }
-    return styleValuesAreEqualForReuse(style, other.style)
-  }
+extension ConcreteStyleBox: AnyProgressViewStyleBox where S: ProgressViewStyle {
 
   @MainActor
   func resolveBody(configuration: ProgressViewStyleConfiguration, in context: ResolveContext)
     -> ResolvedNode
   {
-    resolveStyleBody(
-      bindingForwardedDynamicPropertyCaptures(style).makeBody(configuration: configuration),
-      styleLabel: style.snapshotLabel, in: context)
+    resolveBody(
+      configuration: configuration, styleLabel: style.snapshotLabel, in: context,
+      makeBody: { style, configuration in style.makeBody(configuration: configuration) })
   }
 }